New class IsValid.cs, Moved Method IsValidPersonNumber to class
